Add files via upload
Tired of going online to convert any word file to pdf...
Here's a simple solution. Run the file from the Windows command line with the first argument as your word file and the second one as the name you want to assign to the new pdf file.
And booom...your file is converted with a "Converted Successfully" Message on the screen.

If you get any error, feel free to comment on the post.